# zema_store_backend/src/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zema Store API with Swagger",
        "version": "1.0.0",
        "description": "Zema Store API",
        "contact": {
            "name": "Zema Store",
            "url": "http://zema-store.herokuapp.com",
            "email": "henokadane8@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://zema-store.herokuapp.com/api"
        },
        {
            "url": "http://localhost:3000/api"
        }
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/songs/artist/{artistId}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Returns list of songs",
                "description": "Returns list of songs that belogn to an artist",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "artistId",
                        "required": true,
                        "description": "id of the specified artist"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "type": "integer",
                        "description": "The page number of song items to return"
                    },
                    {
                        "in": "query",
                        "name": "sortBy",
                        "type": "string",
                        "description": "song object field to sort returned songs",
                        "enum": [
                            "title",
                            "releaseDate",
                            "createdAt"
                        ],
                        "example": "title:asc"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Returns list of artists",
                "description": "Returns list of all artists",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "type": "integer",
                        "description": "The page number of artist items to return"
                    },
                    {
                        "in": "query",
                        "name": "sortBy",
                        "type": "string",
                        "description": "artist object field to sort returned artists",
                        "enum": [
                            "createdAt"
                        ],
                        "example": "createdAt:desc"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Adds an artist",
                "description": "Adds an artist",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Artist"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/artists/{id}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Returns artist data",
                "description": "Returns artist data specified by id",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "id of the specified artist"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Artists"
                ],
                "summary": "Updates artist data",
                "description": "Updates artist data with specified id",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "id of the specified artist"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Artist"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Deletes artist data",
                "description": "Deletes artist data with specified id",
                "security": [
                    {
                        "bearerAurhorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "id of the specified artist"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/albums/artist/{artistId}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Returns list of albums",
                "description": "Returns list of albums that belong to an artist",
                "security": [
                    {
                        "bearerAuthorization": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "artistId",
                        "required": true,
                        "description": "id of the specified album"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "type": "integer",
                        "description": "The page number of album items to return"
                    },
                    {
                        "in": "query",
                        "name": "sortBy",
                        "type": "string",
                        "description": "album object field to sort returned albums",
                        "enum": [
                            "title",
                            "createdAt"
                        ],
                        "example": "title:asc"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OkResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuthorization": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token should be accessToken for user"
            }
        },
...
        "schemas": {
            "SuccessMessage": {
                "type": "string",
                "description": "Success message",
                "nullable": true,
                "required": true
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message",
                        "required": true
                    },
                    "data": {
                        "type": "object",
                        "required": true,
                        "nullable": true
                    }
                }
            },
            "OkResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "$ref": "#/components/schemas/SuccessMessage"
                    },
                    "data": {
                        "type": "object",
                        "required": true,
                        "nullable": false
                    }
                },
                "required": [
                    "message",
                    "data"
                ]
            },
...
            "Artist": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "Valid email address"
                    },
                    "password": {
                        "type": "string",
                        "description": "Password"
                    },
                    "photo": {
                        "type": "string",
                        "format": "binary",
                        "description": "Image of the artist"
                    },
                    "phone": {
                        "type": "string",
                        "description": "User phone number"
                    },
                    "fullName": {
                        "type": "string",
                        "description": "Full Name"
                    }
                },
                "required": [
                    "email",
                    "password",
                    "fullName",
                    "address"
                ]
            },
...
        }
    }
}